requirements.txt
pandas
flask
pymongo
requests
python-dotenv
openpyxl

mysql-connector-python